Fix the redesigned burstsort so it does not sort the null bucket entries which is a waste of cycles (consumed strings do not need to be sorted).
Fixed the redesigned burstsort to break up the copying of the null buckets
into separate jobs so they can be processed in parallel, otherwise you may
have a large number of elements being processed on a single thread and
nothing being done on the other threads.

Added many more unit tests to test all of this new code.